requirements.txt
GitPython
Pillow
